# PR: Optimize downloader