I ran multiple experiments to find the optimal settings for the downloader. 

For any number of threads, a 4MB chunk size is optimal, so I updated the default:

```
Download benchmarks (10GB remote, single file local)

Threads  Chunk Size  Elapsed    Rate
-------  ----------  ---------  ------------
      4         4MB  224.3178s  45.6497 MB/s
      8        32MB  251.6219s  40.6960 MB/s
      8        16MB  237.1416s  43.1810 MB/s
      8         8MB  225.4522s  45.4198 MB/s
      8         4MB  203.5102s  50.3169 MB/s
      8         2MB  224.7242s  45.5670 MB/s
      8         1MB  354.7242s  28.8675 MB/s
     16        64MB  332.2181s  30.8231 MB/s
     16         8MB  261.7113s  39.1271 MB/s
     16         4MB  236.5254s  43.2934 MB/s
```
```
Download benchmarks (1GB remote, single file local)

Threads  Chunk Size  Elapsed    Rate
-------  ----------  ---------  ------------
     16        64MB   64.6512s  15.8388 MB/s
     16         4MB   10.7376s  95.3658 MB/s
      8        64MB   40.0978s  25.5376 MB/s
      8         4MB   13.8160s  74.1170 MB/s
```
I also updated the upload/download CLI commands, which I modified for the benchmarks.